fix bugssssss = =!

+++ launch/autopilot.sh
@@ -32,24 +32,17 @@ MOGO_LOG() {
     local code=$1
     local msg=$2
     local append=$(cat $ABS_PATH/../config/mogo_report_msg.json | python -c "import json,sys; sys.stdout.write(json.dumps(json.load(sys.stdin).get('$code')));sys.stdout.write('\n');" | sed 's/[\{\}]//g')
-    if [ -z "$append" ];then
-        if [ "$(echo $code | cut -c 1)" == 'E' ];then
+    if [ -z "$append" ]; then
+        if [ "$(echo $code | cut -c 1)" == 'E' ]; then
             level="error"
-        elif [ "$(echo $code | cut -c 1)" == 'I' ];then
+        elif [ "$(echo $code | cut -c 1)" == 'I' ]; then
             level="info"
         fi
         append="\"level\":\"$level\", \"code\":\"$code\""
     fi
-    echo "{\"timestamp\": {\"sec\": $(date +"%s"), \"nsec\": $(date +"%N")}, \"src\": \"$this\", $append, \"msg\": \"$msg\"}" >> $MOGOLOGFILE
+    echo "{\"timestamp\": {\"sec\": $(date +"%s"), \"nsec\": $(date +"%N")}, \"src\": \"$this\", $append, \"msg\": \"$msg\"}" >>$MOGOLOGFILE
 }
 
-kill_ros() {
-    core_pid=$(ps aux | grep -v grep | grep -w "rosmaster --core" | awk '{print $2}')
-    if [ ! -z "$core_pid" ]; then
-        rosnode kill -a
-        kill $core_pid
-    fi
-}
 set_bashrc() {
     HOSTNAME="export ROS_HOSTNAME=$ros_machine"
     MASTER_URI="export ROS_MASTER_URI=http://$ros_master:11311"
@@ -147,13 +140,18 @@ log4j.appender.${ErrorAppender}.layout.ConversionPattern=[%-5p] %d{yyyy-MM-dd HH
 }
 
 get_all_launch_files() {
-    if [ ! -f $ListFile -o $(wc -l $ListFile| awk '{print $1}') -eq 0 ];then
+    local file_set
+    if [ ! -f $ListFile -o $(sed '/^\s*$/d' $ListFile | wc -m) -eq 0 ]; then
         MOGO_LOG "EINIT_BOOT" "$ListFile doesn't exist or it's empty"
         LoggingERR "$ListFile doesn't exist or it's empty"
         return
     fi
     if [ -n "$opt_onenode" ]; then
         launch_line=$(grep -w $opt_onenode $ListFile)
+        if [ -z "$launch_line" ];then
+            LoggingERR "cannot find \"$opt_onenode\" in $ListFile"
+            exit 1
+        fi
         include_files=$(roslaunch --files $launch_line 2>/dev/null)
         if [ $? -ne 0 ]; then
             LoggingERR "cannot find $launch_line"
@@ -161,14 +159,21 @@ get_all_launch_files() {
         fi
         for child_file in $include_files; do
             [[ -z "$child_file" ]] && continue
-            pkg_name=$(xmllint --xpath "//@pkg" $child_file 2>/dev/null | sed 's/\"//g')
-            [[ -z "$pkg_name" ]] && continue
+            [[ $(echo $file_set | grep -w -c "$child_file:") -ne 0 ]] && continue
+            type_name=$(xmllint --xpath "//node[@pkg!='rviz']/@type" $child_file 2>/dev/null | sed 's/\"//g')
+            [[ -z "$type_name" ]] && continue
+            map_restart_times[$child_file]=-1
             launch_files_array[$arr_idx]="$child_file"
             arr_idx=$((arr_idx + 1))
+            for f in $(roslaunch --files $child_file);do
+                [[ $(echo $file_set | grep -w -c "$f:") -ne 0 ]] && continue
+                file_set="$f:$file_set"
+            done
         done
         return
     fi
     while read launch_file || [[ -n $launch_file ]]; do
+        [[ -z "$launch_file" ]] && continue
         include_files=$(roslaunch --files $launch_file 2>/dev/null)
         if [ $? -ne 0 ]; then
             LoggingERR "cannot find $launch_file"
@@ -177,10 +182,17 @@ get_all_launch_files() {
         fi
         for child_file in $include_files; do
             [[ -z "$child_file" ]] && continue
-            pkg_name=$(xmllint --xpath "//@pkg" $child_file 2>/dev/null | sed 's/\"//g')
-            [[ -z "$pkg_name" ]] && continue
+            [[ $(echo $file_set | grep -w -c "$child_file:") -ne 0 ]] && continue
+            type_name=$(xmllint --xpath "//node[@pkg!='rviz']/@type" $child_file 2>/dev/null | sed 's/\"//g')
+            [[ -z "$type_name" ]] && continue
+            map_restart_times[$child_file]=-1
             launch_files_array[$arr_idx]="$child_file"
             arr_idx=$((arr_idx + 1))
+            # 可能存在的问题:当某launch文件中既包含节点又包含include时，如果roslaunch --files命令先打印子file,则还会出现启动多次的问题，待修复
+            for f in $(roslaunch --files $child_file);do
+                [[ $(echo $file_set | grep -w -c "$f:") -ne 0 ]] && continue
+                file_set="$f:$file_set"
+            done
         done
     done <$ListFile
 }
@@ -263,83 +275,84 @@ keep_alive() {
                         start_onenode "$launch_file"
                         break
                     # try to restart failed
-                    elif [ ${map_restart_times[$launch_file]} -eq 6 ];then
+                    elif [ ${map_restart_times[$launch_file]} -eq 6 ]; then
                         map_restart_times[$launch_file]=999
                         MOGO_LOG "EMAP_NODE_DEAD" "restart $t failed 5 times,cancel to restart"
                         LoggingERR "restart $t failed 5 times,cancel to restart"
                     # launch error
-                    elif [ ${map_restart_times[$launch_file]} -eq 0 ];then
+                    elif [ ${map_restart_times[$launch_file]} -eq 0 ]; then
                         MOGO_LOG "EINIT_BOOT" "launch $t failed"
                         LoggingERR "launch $t failed"
-                        map_restart_times[$launch_file]=1
+                        # map_restart_times[$launch_file]=1 #bug:如果launch里面有多个node，则只打印一次日志
                     fi
                 else
                     if [ ${map_restart_times[$launch_file]} -gt 1 -a ${map_restart_times[$launch_file]} -lt 5 ]; then
                         MOGO_LOG "IBOOT_LOCAL_NODES_STATUS" "restart $t succeed"
                         LoggingINFO "$t restart succeed"
-                    elif [ ${map_restart_times[$launch_file]} -eq 0 ];then
+                    elif [ ${map_restart_times[$launch_file]} -eq 0 ]; then
                         MOGO_LOG "IBOOT_LOCAL_NODES_STATUS" "launch $t succeed"
-                        LoggingINFO "launch $t succeed"
-                        map_restart_times[$launch_file]=1
+                        LoggingINFO "launch $t succeed" 
+                        # map_restart_times[$launch_file]=1 #bug:如果launch里面有多个node，则只打印一次日志
                     fi
                     #pr
-                    pid=$(ps -ef | grep "__name:=$t" | grep -v grep | awk '{print $2}')  
-                    if [ -z "$pid" ];then
+                    pid=$(ps -ef | grep "__name:=$t" | grep -v grep | awk '{print $2}')
+                    if [ -z "$pid" ]; then
                         continue
                     fi
                     priority=$(top -b -n 1 -p $pid | grep $pid | awk '{print $(NF-9)}')
-                    if [ "$priority" == "rt" ];then
+                    if [ "$priority" == "rt" ]; then
                         continue
                     fi
 
-					case "$t" in
-					"DongFeng_E70_can_adapter" | "jinlv_can_adapter" | "hongqih9_can_adapter")
-						if [ $priority -gt 0 ];then
-							chrt -p -r 99 $pid
-						fi
-						;;
-					"controller")
-						if [ $priority -gt 0 ];then
+                    case "$t" in
+                    "DongFeng_E70_can_adapter" | "jinlv_can_adapter" | "hongqih9_can_adapter")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 99 $pid
                         fi
                         ;;
-						
-					"localization" | "drivers_gnss" | "drivers_gnss_zy")
-						if [ $priority -gt 0 ];then
+                    "controller")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 99 $pid
                         fi
                         ;;
-					"local_planning")
-						if [ $priority -gt 0 ];then
+
+                    "localization" | "drivers_gnss" | "drivers_gnss_zy")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 99 $pid
                         fi
                         ;;
-					"hadmap_server" | "hadmap_engine_node")
-						if [ $priority -gt 0 ];then
+                    "local_planning")
+                        if [ $priority -gt 0 ]; then
+                            chrt -p -r 99 $pid
+                        fi
+                        ;;
+                    "hadmap_server" | "hadmap_engine_node")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 80 $pid
                         fi
                         ;;
-					"perception_fusion2" | "perception_fusion")
-						if [ $priority -gt 0 ];then
+                    "perception_fusion2" | "perception_fusion")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 79 $pid
                         fi
                         ;;
-					"rs_perception_node" | "trt_yolov5")	
-						if [ $priority -gt 0 ];then
+                    "rs_perception_node" | "trt_yolov5")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 69 $pid
                         fi
                         ;;
-					"drivers_camera_sensing60" | "drivers_camera_sensing30" | "drivers_camera_sensing120" | "drivers_robosense_node")
-						if [ $priority -gt 0 ];then
+                    "drivers_camera_sensing60" | "drivers_camera_sensing30" | "drivers_camera_sensing120" | "drivers_robosense_node")
+                        if [ $priority -gt 0 ]; then
                             chrt -p -r 59 $pid
                         fi
                         ;;
-					*)
-						;;
-					esac
+                    *) ;;
+
+                    esac
                 fi
             done
             [[ $proc_stat -eq 0 ]] && map_restart_times[$launch_file]=1
+            [[ $proc_stat -ne 0 && ${map_restart_times[$launch_file]} -eq 0 ]] && map_restart_times[$launch_file]=1
         done
     done
 }
@@ -358,7 +371,7 @@ LoggingERR() {
 }
 
 install_ros_log() {
-    [[ ! -d /autocar-code/install ]] && return 
+    [[ ! -d /autocar-code/install ]] && return
     src_so_path=$(find /autocar-code/install/ -name 'libroscpp.so' | head -n 1)
     dst_so_path=$(find /opt -name 'libroscpp.so' | head -n 1)
     conf_path=$(find /autocar-code/install/ -name 'ros_statics.conf' | head -n 1)
@@ -370,19 +383,29 @@ install_ros_log() {
     \cp -rf $conf_path $dst_conf_path
 }
 
-add_privilege_monitor_gnss(){
-     rm  /home/mogo/data/log/location.txt  -rf
-     rm  /home/mogo/data/log/msg_info_log.txt  -rf
-     rm  /home/mogo/data/log/msg_error_log.txt  -rf
-     rm  /home/mogo/data/log/topic_hz_log.txt  -rf
-     chmod -R  777 /autocar-code/install/share/monitor_gnss
-     chmod -R  777 /autocar-code/install/share/monitor_collect
+add_privilege_monitor_gnss() {
+    [[ ! -d /autocar-code/install ]] && return
+    rm /home/mogo/data/log/location.txt -rf
+    rm /home/mogo/data/log/msg_info_log.txt -rf
+    rm /home/mogo/data/log/msg_error_log.txt -rf
+    rm /home/mogo/data/log/topic_hz_log.txt -rf
+    chmod -R 777 /autocar-code/install/share/monitor_gnss >/dev/null 2>&1
+    chmod -R 777 /autocar-code/install/share/monitor_collect >/dev/null 2>&1
 }
 
 start_core() {
+    [[ -n "$opt_onenode" ]] && return
     roscore 2>&1 >$ROS_LOG_DIR/roscore.log &
     roscore_pid=$!
     chrt -p -r 99 $roscore_pid
+}
+kill_ros() {
+    [[ -n "$opt_onenode" ]] && return
+    core_pid=$(ps aux | grep -v grep | grep -w "rosmaster --core" | awk '{print $2}')
+    if [ ! -z "$core_pid" ]; then
+        rosnode kill -a
+        kill $core_pid
+    fi
 }
 
 _exit() {
@@ -402,7 +425,7 @@ declare -g opt_onenode
 declare -g opt_alive=1
 declare -g launch_files_array
 declare -g arr_idx=0
-declare roscore_pid
+declare -g roscore_pid
 declare -g LOGFILE
 declare -g ERRFILE
 declare -g LOG_DIR="/home/mogo/data/log"
@@ -439,6 +462,31 @@ while true; do
         ;;
     esac
 done
+
+#check system time
+export last_launch_time
+if [ -f $ABS_PATH/launch_time ]; then
+    last_launch_time=$(cat $ABS_PATH/launch_time)
+fi
+if [ -z "$last_launch_time" ]; then
+    last_launch_time="20211222211202" #随便初始的一个日期时间，不要在意细节
+fi
+while [ true ]; do
+    systime=$(date +"%Y%m%d%H%M%S")
+    if [ $systime -gt $last_launch_time ]; then
+        MOGO_LOG "IINIT_TIME_SYNC" "systime synchronization at $systime"
+        echo $systime >$ABS_PATH/launch_time
+        break
+    fi
+    sleep 1
+done
+curtime=$(date +"%Y%m%d%H%M%S")
+LOGFILE="/home/mogo/data/log/autopilot-${curtime}.log"
+ERRFILE="/home/mogo/data/log/autopilot-${curtime}.err"
+LoggingINFO "systime synchronization at $curtime"
+find ${LOG_DIR} -maxdepth 1 -mtime +3 -type d -exec rm -Rf {} \;
+find ${LOG_DIR} -name "autopilot-*" -type f -mtime +1 -exec rm -Rf {} \;
+find $BAG_DIR -maxdepth 1 -mtime +1 -type d -exec rm -Rf {} \;
 
 [[ ! -d $MOGO_LOG_DIR ]] && mkdir -p $MOGO_LOG_DIR
 [ ! -f $MOGO_MSG_CONFIG ] && MOGO_LOG "EINIT_BOOT" "cannot get mogo_msg_config,report could be incompletion"
@@ -499,8 +547,8 @@ export xavier_type="single"        #单xavier or 双xavier
 export ros_machine="${ros_master}" #主机:rosmaster 从机:rosslave
 # 判断是否为双Xavier
 ethnet_ip=$(ifconfig | grep -v "inet6" | grep -Eo '192[.]168([.][0-9]+){2}' | grep -v "255")
-if [ -z "$ethnet_ip" ];then
-    LoggingERR "ip address is null" 
+if [ -z "$ethnet_ip" ]; then
+    LoggingERR "ip address is null"
     MOGO_LOG "EHW_NET" "ip address is null"
     exit 1
 fi
@@ -557,64 +605,42 @@ export OMP_NUM_THREADS=1
 export BASHRC="source ${SETUP_ROS} && source ${SETUP_AUTOPILOT}"
 source $SETUP_ROS
 source $SETUP_AUTOPILOT
+[[ ! -d $BAG_DIR ]] && mkdir -p $BAG_DIR
 
-get_all_launch_files
+get_all_launch_files #获取所有需要启动的launch文件
 if [ -n "$opt_onenode" ]; then
     ROS_LOG_DIR=$(readlink ${LOG_DIR}/latest)
+    [[ -z "ROS_LOG_DIR" ]] && ROS_LOG_DIR="${LOG_DIR}/$(date +"%Y%m%d_%H%M%S")"
+    [[ ! -d $ROS_LOG_DIR ]] && mkdir -p $ROS_LOG_DIR
+    ln -snf $ROS_LOG_DIR ${LOG_DIR}/latest
     start_node
     sleep 1
     exit 0
-fi
-
-install_ros_log
-add_privilege_monitor_gnss
-
-#check system time
-export last_launch_time
-if [ -f $ABS_PATH/launch_time ]; then
-    last_launch_time=$(cat $ABS_PATH/launch_time)
-fi
-if [ -z "$last_launch_time" ]; then
-    last_launch_time="20211222211202" #随便初始的一个日期时间，不要在意细节
-fi
-while [ true ]; do
-    systime=$(date +"%Y%m%d%H%M%S")
-    if [ $systime -gt $last_launch_time ]; then
-        LoggingINFO "systime synchronization at $systime"
-        MOGO_LOG "IINIT_TIME_SYNC" "systime synchronization at $systime"
-        echo $systime >$ABS_PATH/launch_time
+else
+    # 自动驾驶自检
+    stat_file="/home/mogo/data/vehicle_monitor/check_system.txt"
+    while [ true ]; do
+        # 运维自检状态
+        sleep 1
+        if [ -f $stat_file ]; then
+            [[ $(head -1 $stat_file) -ne 0 ]] && LoggingERR "system check failed" && continue
+        fi
+        # python $ABS_PATH/mogodoctor.py c>>$LOGFILE 2>>$ERRFILE
+        # if [ $? -eq 0 ]; then
+        #     break
+        # fi
+        # LoggingERR "autopilot check failed"
         break
-    fi
-    sleep 1
-done
-curtime=$(date +"%Y%m%d%H%M%S")
-[[ -f $LOGFILE ]] && mv $LOGFILE "/home/mogo/data/log/autopilot-${curtime}.log"
-[[ -f $ERRFILE ]] && mv $ERRFILE "/home/mogo/data/log/autopilot-${curtime}.err"
-LOGFILE="/home/mogo/data/log/autopilot-${curtime}.log"
-ERRFILE="/home/mogo/data/log/autopilot-${curtime}.err"
+    done
+    python $ABS_PATH/mogodoctor.py c >>$LOGFILE 2>>$ERRFILE
+    install_ros_log
+    add_privilege_monitor_gnss
+    ROS_LOG_DIR="${LOG_DIR}/$(date +"%Y%m%d_%H%M%S")"
 
-# 自动驾驶自检
-stat_file="/home/mogo/data/vehicle_monitor/check_system.txt"
-while [ true ]; do
-    # 运维自检状态
-    sleep 1
-    if [ -f $stat_file ]; then
-        [[ $(head -1 $stat_file) -ne 0 ]] && LoggingERR "system check failed" && continue
-    fi
-    # python $ABS_PATH/mogodoctor.py c>>$LOGFILE 2>>$ERRFILE
-    # if [ $? -eq 0 ]; then
-    #     break
-    # fi
-    # LoggingERR "autopilot check failed"
-    break
-done
-python $ABS_PATH/mogodoctor.py c>>$LOGFILE 2>>$ERRFILE
-
-ROS_LOG_DIR="${LOG_DIR}/$(date +"%Y%m%d_%H%M%S")"
+fi
 [[ ! -d $ROS_LOG_DIR ]] && mkdir -p $ROS_LOG_DIR
 ln -snf $ROS_LOG_DIR ${LOG_DIR}/latest
 LoggingINFO "ROS_LOG_DIR=$ROS_LOG_DIR"
-[[ ! -d $BAG_DIR ]] && mkdir -p $BAG_DIR
 export LOG_ENV="export GLOG_logtostderr=1; export GLOG_colorlogtostderr=1; export ROS_LOG_DIR=${ROS_LOG_DIR}; export ROS_MASTER_URI=http://${ros_master}:11311; export ROS_HOSTNAME=${ros_machine}"
 
 LoggingINFO "path : $ABS_PATH"
@@ -626,18 +652,14 @@ if [ $# -eq 0 ]; then
 fi
 LoggingINFO "command : $0 $@"
 
-find ${LOG_DIR} -maxdepth 1 -mtime +3 -type d -exec rm -Rf {} \;
-find ${LOG_DIR} -name "autopilot-*" -type f -mtime +1 -exec rm -Rf {} \;
-find $BAG_DIR -maxdepth 1 -mtime +1 -type d -exec rm -Rf {} \;
-
 kill_ros
-
 if [ "$ros_machine" == "$ros_master" ]; then
     start_core
 else
     while [ true ]; do
-        timeout 3 cat </dev/tcp/${ros_master}/11311
+        timeout 1 cat </dev/tcp/${ros_master}/11311
         if [ $? -ne 124 ]; then
+            sleep 2
             continue
         fi
         break
@@ -658,7 +680,7 @@ ret=$?
 if [ $ret -eq 124 ]; then
     LoggingERR "update config timeout"
     MOGO_LOG "EMAP_NODE" "update config timeout"
-elif [ $ret -eq 0 ];then
+elif [ $ret -eq 0 ]; then
     MOGO_LOG "IBOOT_CONFIG_UPDATE" "update config finished"
     LoggingINFO "update config finished"
 fi
